Move label.go file from package testsuite to util
This was done in order to be able to move functions from tests/utils
to libstorage package without having a circular dependencies
between the testsuite and the libstorage packages.
