+++ tests/util/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["util.go"],
+    srcs = [
+        "labels.go",
+        "util.go",
+    ],
     importpath = "kubevirt.io/kubevirt/tests/util",
     visibility = ["//visibility:public"],
     deps = [
