+++ tests/vmi_lifecycle_test.go
@@ -474,7 +474,7 @@ var _ = Describe("[rfe_id:273][crit:high][arm64][vendor:cnv-qe@redhat.com][level
 							Name:      "nonexistent",
 							Namespace: vmi.Namespace,
 							Labels: map[string]string{
-								testsuite.SecretLabel: "nonexistent",
+								util.SecretLabel: "nonexistent",
 							},
 						},
 						Type: "Opaque",
